review: add pi fixture (registry pi branch exercised); split codex fixtures for max-lines
