+++ src/main/ai-vault/session-scanner-incremental-fixtures.ts
@@ -1,5 +1,7 @@
 import type { AiVaultAgent } from '../../shared/ai-vault-types'
 
+import { codexFixture } from './session-scanner-codex-fixtures'
+
 // Line builders for the incremental-parse differential tests: each agent gets
 // a seed transcript, an appended continuation, and a truncated rewrite, all in
 // that agent's real on-disk JSONL record shapes.
@@ -11,94 +13,6 @@ export type IncrementalAgentFixture = {
   appendLines: string[]
   truncatedLines: string[]
 }
-
-const CODEX_SESSION_ID = '019f0000-1111-7222-8333-444444444444'
-
-function codexLine(record: Record<string, unknown>): string {
-  return JSON.stringify(record)
-}
-
-export function codexFixture(): IncrementalAgentFixture {
-  return {
-    agent: 'codex',
-    fileName: `rollout-2026-05-01T10-00-00-${CODEX_SESSION_ID}.jsonl`,
-    seedLines: [
-      codexLine({
-        timestamp: '2026-05-01T10:00:00.000Z',
-        type: 'session_meta',
-        payload: { id: CODEX_SESSION_ID, cwd: '/repo/app', git: { branch: 'feature/vault' } }
-      }),
-      codexLine({
-        timestamp: '2026-05-01T10:00:05.000Z',
-        type: 'response_item',
-        payload: { type: 'message', role: 'user', content: 'codex seed question' }
-      }),
-      codexLine({
-        timestamp: '2026-05-01T10:00:10.000Z',
-        type: 'response_item',
-        payload: { type: 'message', role: 'assistant', content: 'codex seed answer' }
-      }),
-      codexLine({
-        timestamp: '2026-05-01T10:00:11.000Z',
-        type: 'event_msg',
-        payload: {
-          type: 'token_count',
-          info: { total_token_usage: { input_tokens: 100, output_tokens: 40, total_tokens: 140 } },
-          model: 'gpt-5.1-codex'
-        }
-      })
-    ],
-    appendLines: [
-      codexLine({
-        timestamp: '2026-05-01T10:05:00.000Z',
-        type: 'event_msg',
-        payload: { type: 'user_message', message: 'codex follow-up' }
-      }),
-      codexLine({
-        timestamp: '2026-05-01T10:05:20.000Z',
-        type: 'event_msg',
-        payload: { type: 'agent_message', message: 'codex incremental answer' }
-      }),
-      codexLine({
-        timestamp: '2026-05-01T10:05:21.000Z',
-        type: 'event_msg',
-        payload: {
-          type: 'token_count',
-          info: { total_token_usage: { input_tokens: 220, output_tokens: 90, total_tokens: 310 } }
-        }
-      })
-    ],
-    truncatedLines: [
-      codexLine({
-        timestamp: '2026-05-01T10:00:00.000Z',
-        type: 'session_meta',
-        payload: { id: CODEX_SESSION_ID, cwd: '/repo/app' }
-      }),
-      codexLine({
-        timestamp: '2026-05-01T10:00:05.000Z',
-        type: 'event_msg',
-        payload: { type: 'user_message', message: 'rewritten only turn' }
-      })
-    ]
-  }
-}
-
-export function codexWorkerFixtureLines(): string[] {
-  return [
-    codexLine({
-      timestamp: '2026-05-01T10:00:00.000Z',
-      type: 'session_meta',
-      payload: { id: CODEX_SESSION_ID, cwd: '/repo/app', thread_source: 'subagent' }
-    }),
-    codexLine({
-      timestamp: '2026-05-01T10:00:05.000Z',
-      type: 'event_msg',
-      payload: { type: 'user_message', message: 'worker turn' }
-    })
-  ]
-}
-
-export const CODEX_FIXTURE_SESSION_ID = CODEX_SESSION_ID
 
 export function cursorFixture(): IncrementalAgentFixture {
   const line = (role: string, text: string, at: string) =>
@@ -233,6 +147,47 @@ export function openclawFixture(): IncrementalAgentFixture {
   }
 }
 
+// Pi shares OpenClaw's message-graph format and factory, but gets its own
+// fixture so the registry's 'pi' branch is exercised explicitly.
+export function piFixture(): IncrementalAgentFixture {
+  return {
+    agent: 'pi',
+    fileName: 'bbbbbbbb-cccc-4ddd-8eee-ffffffffffff.jsonl',
+    seedLines: [
+      JSON.stringify({
+        type: 'session',
+        id: 'pi-session-1',
+        cwd: '/repo/app',
+        timestamp: '2026-05-01T10:00:00.000Z'
+      }),
+      JSON.stringify({
+        type: 'message',
+        message: { role: 'user', content: 'pi seed question' },
+        timestamp: '2026-05-01T10:00:05.000Z'
+      })
+    ],
+    appendLines: [
+      JSON.stringify({
+        type: 'model_change',
+        modelId: 'pi-2',
+        timestamp: '2026-05-01T10:00:30.000Z'
+      }),
+      JSON.stringify({
+        type: 'message',
+        message: {
+          role: 'assistant',
+          content: 'pi incremental answer',
+          usage: { input_tokens: 30, output_tokens: 10 }
+        },
+        timestamp: '2026-05-01T10:01:00.000Z'
+      })
+    ],
+    truncatedLines: [
+      JSON.stringify({ type: 'session', id: 'pi-session-1', timestamp: '2026-05-01T10:00:00.000Z' })
+    ]
+  }
+}
+
 export function geminiJsonlFixture(): IncrementalAgentFixture {
   return {
     agent: 'gemini',
@@ -279,6 +234,7 @@ export function allIncrementalAgentFixtures(): IncrementalAgentFixture[] {
     copilotFixture(),
     droidFixture(),
     openclawFixture(),
+    piFixture(),
     geminiJsonlFixture()
   ]
 }
